editorconfig: indent text files with tabs
Previously, the .editorconfig did not specify an indentation style for
text files.  However, a quick look for indentation-like spacing suggest
that tabs are more common for documentation:

$ git grep -Pe '^ {4}' -- '*.txt' |wc -l
2683
$ git grep -Pe '^\t' -- '*.txt' |wc -l
14011

Note that there are a lot of files that indent list continuations (and
other things) with a single space -- if the first search was made
without the fixed quantifier the result would look very different.
However, the result does correspond with my anecdotal experience when
editing git documentation.

This commit adds *.txt to .editorconfig as an extension that should be
indented with tabs.

Signed-off-by: Hans Jerry Illikainen <hji@dyntopia.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>